fix(onehub): move parseYouTubeId out of the route so the build passes
A Next.js App Router route file may only export the HTTP verbs and a fixed
set of config fields. `parseYouTubeId` was exported from
app/api/onehub/videos/route.ts purely so the test could import it, which
fails `next build`:

  Type error: Route "app/api/onehub/videos/route.ts" does not match the
  required types of a Next.js Route.
    "parseYouTubeId" is not a valid Route export field.

CI never caught this because it runs typecheck, lint and test but not
`next build` — the rule lives in the .next/types files Next generates
during the build, so `tsc --noEmit` is happy. The Vercel deploy of #128
failed on it, which is why the quotation PDF never reached production.

The helper moves to src/lib/youtube-id.ts unchanged; the route and the
test import it from there.
