+++ apps/web/app/api/onehub/videos/route.ts
@@ -5,23 +5,7 @@ import { z } from "zod";
 import { success, error, parseBody } from "@/lib/api-utils";
 import { requireAuth, AuthError } from "@/lib/api-helpers";
 import { supabaseAdmin } from "@/lib/supabase-admin";
-
-/** Accepts a bare id, a watch?v= link, a youtu.be link or an embed link. */
-export function parseYouTubeId(input: string): string | null {
-  const value = input.trim();
-  if (/^[A-Za-z0-9_-]{11}$/.test(value)) return value;
-  const patterns = [
-    /[?&]v=([A-Za-z0-9_-]{11})/,
-    /youtu\.be\/([A-Za-z0-9_-]{11})/,
-    /\/embed\/([A-Za-z0-9_-]{11})/,
-    /\/shorts\/([A-Za-z0-9_-]{11})/,
-  ];
-  for (const re of patterns) {
-    const m = value.match(re);
-    if (m) return m[1];
-  }
-  return null;
-}
+import { parseYouTubeId } from "@/lib/youtube-id";
 
 const videoSchema = z.object({
   id: z.string().uuid().optional(),
